Extract a new class that is responsible for fetching flags (polling mechanism)
